python/requirements-trocr.txt
accelerate==0.28.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
datasets==2.18.0
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.2.2
Faker==24.7.1
filelock==3.18.0
fire==0.7.0
Flask==3.0.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.3.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==3.0.0
pandas==2.2.3
pdf2image==1.17.0
pillow==10.2.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyclipper==1.3.0.post6
pydantic==2.11.0
pydantic_core==2.33.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
shapely==2.0.7
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
sympy==1.13.3
termcolor==2.5.0
tifffile==2025.3.13
tokenizers==0.15.2
torch==2.1.2
tqdm==4.66.2
transformers==4.39.3
typer==0.12.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.18.3

# Für Matplotlib-Vorschau via TkAgg notwendig:
# Stelle sicher, dass tkinter installiert ist (z.B. über python.org-Installer)